requirements.txt
fastapi
pydantic
uvicorn
requests
python-dotenv